requirements.txt
streamlit
matplotlib.pyplot
seaborn
fastf1
numpy
gnews
datetime